+++ web/src/components/hostdashboard/HostListings.js
@@ -5,26 +5,13 @@ import add from "../../images/icons/host-add.png";
 import { useNavigate } from 'react-router-dom';
 import { Auth } from "aws-amplify";
 import spinner from "../../images/spinnner.gif";
-import ImageSlider from "./ImageSlider";
+import PageSwitcher from "../utils/PageSwitcher";
 
 function HostListings() {
     const [accommodations, setAccommodations] = useState([]);
     const [isLoading, setIsLoading] = useState(true);
     const [userId, setUserId] = useState(null);
     const navigate = useNavigate();
-    const formatDate = (dateString) => {
-        const date = new Date(dateString);
-        const day = date.getDate().toString().padStart(2, '0');
-        const month = (date.getMonth() + 1).toString().padStart(2, '0');
-        const year = date.getFullYear();
-
-        return `${day}/${month}/${year}`;
-    }
-    const getFeatures = (features) => {
-        return Object.entries(features)
-            .filter(([key, value]) => value === true)
-            .map(([key, value]) => key);
-    }
     useEffect(() => {
         const setUserIdAsync = async () => {
             try {
@@ -104,36 +91,7 @@ function HostListings() {
                                     <img src={spinner}/>
                                 </div>
                             ) : accommodations.length > 0 ? (
-                                accommodations.map((accommodation) => (
-                                    <div key={accommodation.ID} className="accommodation-tab">
-                                        <div className="accommodation-left">
-                                            <p className="accommodation-title">{accommodation.Title}</p>
-                                            <p>{accommodation.Country},
-                                                {accommodation.City},
-                                                {accommodation.Street},
-                                                {accommodation.PostalCode}
-                                            </p>
-                                            <ImageSlider images={accommodation.Images} />
-                                        </div>
-                                        <div className="accommodation-right">
-                                        <p>Description: {accommodation.Description}</p>
-                                            <p>Listed on: {formatDate(accommodation.createdAt)}</p>
-                                            <p>Measurements: {accommodation.Measurements}m²</p>
-                                            <p>Features: {accommodation.Features.length > 0 ? (
-                                                getFeatures(accommodation.Features)) : ('none')}
-                                            </p>
-                                            <p>Rent: ${accommodation.Rent}</p>
-                                            { accommodation.StartDate && accommodation.EndDate ?
-                                                (<p>
-                                                    Available from
-                                                    {" " + formatDate(accommodation.StartDate) + " "}
-                                                    to {" " + formatDate(accommodation.EndDate) + " "}
-                                                </p>) :
-                                                (<p>Date range not set</p>)
-                                            }
-                                        </div>
-                                    </div>
-                                ))
+                                <PageSwitcher accommodations={accommodations} amount={5}/>
                             ) : (
                                 <div className="accommodation-box">
                                     <p className="accommodation-alert">It appears that you have not listed any accommodations yet...</p>
